+++ templates/index.html
@@ -6,11 +6,11 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link
       rel="stylesheet"
-      href="{{ url_for('static', filename='/css/reset.css')}}"
+      href="{{url_for('static', filename='/css/reset.css')}}"
     />
     <link
       rel="stylesheet"
-      href="{{ url_for('static', filename='/css/index.css')}}"
+      href="{{url_for('static', filename='/css/index.css')}}"
     />
     <title>Home Page</title>
   </head>
@@ -39,7 +39,8 @@
               placeholder="輸入景點名稱查詢"
             />
             <button type="submit" class="search-btn" id="search-btn">
-              <img src={{ url_for('static', filename='/img/icon_search.png') }}
+              <img
+                src="{{url_for('static', filename='/img/icon_search.png')}}"
               />
             </button>
             <div class="category-list"></div>
@@ -47,13 +48,10 @@
         </div>
       </div>
     </section>
-
     <main></main>
-
     <footer>
       <p class="bold">COPYRIGHT © 2021 台北一日遊</p>
     </footer>
-
     <script src="{{url_for('static', filename='/script/index.js')}}"></script>
   </body>
 </html>
